Split class files
