+++ Xamarin.Forms.Core/PlatformConfiguration/Android/AndroidPlatform.cs
@@ -1,0 +1,5 @@
+
+namespace Xamarin.Forms.PlatformConfiguration.Android
+{
+	public sealed class AndroidPlatform : IConfigPlatform { }
+}
