+++ Xamarin.Forms.Core/PlatformConfiguration/IConfigElement.cs
@@ -1,0 +1,8 @@
+
+namespace Xamarin.Forms.PlatformConfiguration
+{
+	public interface IConfigElement<out T> where T : Element
+	{
+		T Element { get; }
+	}
+}
