+++ Xamarin.Forms.Core/PlatformConfiguration/IConfigPlatform.cs
@@ -1,0 +1,5 @@
+
+namespace Xamarin.Forms.PlatformConfiguration
+{
+	public interface IConfigPlatform { }
+}
