+++ Xamarin.Forms.Core/PlatformConfiguration/IElementConfiguration.cs
@@ -1,0 +1,8 @@
+
+namespace Xamarin.Forms.PlatformConfiguration
+{
+	public interface IElementConfiguration<out TElement> where TElement : Element
+	{
+		IPlatformElementConfiguration<T, TElement> On<T>() where T : IConfigPlatform;
+	}
+}
