+++ Xamarin.Forms.Core/PlatformConfiguration/IPlatformElementConfiguration.cs
@@ -1,0 +1,9 @@
+
+namespace Xamarin.Forms.PlatformConfiguration
+{
+	public interface IPlatformElementConfiguration<out TPlatform, out TElement> : IConfigElement<TElement>
+			where TPlatform : IConfigPlatform
+	 		where TElement : Element
+	{
+	}
+}
