+++ Xamarin.Forms.Core/PlatformConfiguration/Windows/CollapseStyle.cs
@@ -1,0 +1,9 @@
+
+namespace Xamarin.Forms.PlatformConfiguration.Windows
+{
+	public enum CollapseStyle
+	{
+		None,
+		Partial
+	}
+}
