+++ Xamarin.Forms.Core/PlatformConfiguration/Windows/WindowsPlatform.cs
@@ -1,0 +1,5 @@
+
+namespace Xamarin.Forms.PlatformConfiguration.Windows
+{
+	public sealed class WindowsPlatform : IConfigPlatform { }
+}
