+++ Xamarin.Forms.Core/PlatformConfiguration/iOS/iOSPlatform.cs
@@ -1,0 +1,6 @@
+
+namespace Xamarin.Forms.PlatformConfiguration.iOS
+{
+
+	public sealed class iOSPlatform : IConfigPlatform { }
+}
