+++ Xamarin.Forms.Core/Xamarin.Forms.Core.csproj
@@ -87,7 +87,19 @@
     <Compile Include="DateChangedEventArgs.cs" />
     <Compile Include="DelegateLogListener.cs" />
     <Compile Include="EnumerableExtensions.cs" />
-    <Compile Include="PlatformSpecifics.cs" />
+    <Compile Include="PlatformConfiguration\Android\AndroidPlatform.cs" />
+    <Compile Include="PlatformConfiguration\iOS\iOSPlatform.cs" />
+    <Compile Include="PlatformConfiguration\iOS\NavigationPage.cs" />
+    <Compile Include="PlatformConfiguration\Windows\MasterDetailPage.cs" />
+    <Compile Include="PlatformConfiguration\Windows\CollapseStyle.cs" />
+    <Compile Include="PlatformConfiguration\Configuration.cs" />
+    <Compile Include="PlatformConfiguration\IConfigElement.cs" />
+    <Compile Include="PlatformConfiguration\IConfigPlatform.cs" />
+    <Compile Include="PlatformConfiguration\IElementConfiguration.cs" />
+    <Compile Include="PlatformConfiguration\IPlatformElementConfiguration.cs" />
+    <Compile Include="PlatformConfiguration\Android\MasterDetailPage.cs" />
+    <Compile Include="PlatformConfiguration\PlatformConfigurationRegistry.cs" />
+    <Compile Include="PlatformConfiguration\VendorTest.cs" />
     <Compile Include="IFontElement.cs" />
     <Compile Include="DependencyAttribute.cs" />
     <Compile Include="DependencyFetchTarget.cs" />
@@ -196,6 +208,7 @@
     <Compile Include="PanUpdatedEventArgs.cs" />
     <Compile Include="Performance.cs" />
     <Compile Include="PinchGestureUpdatedEventArgs.cs" />
+    <Compile Include="PlatformConfiguration\Windows\WindowsPlatform.cs" />
     <Compile Include="PlatformEffect.cs" />
     <Compile Include="PointTypeConverter.cs" />
     <Compile Include="Internals\PreserveAttribute.cs" />
